fix: cpp core guidelines warning

+++ OpenGL/graphics/obj_loader.cpp
@@ -34,7 +34,7 @@ namespace {
 		if (const auto tokens = string::Split(line, " \t"); tokens.size() == N + 1) {
 			glm::vec<N, T> vec{};
 			for (std::uint8_t i = 0; i < N; ++i) {
-				vec[i] = ParseToken<T>(tokens[i + 1]);
+				vec[i] = ParseToken<T>(tokens[static_cast<std::uint16_t>(i) + 1]);
 			}
 			return vec;
 		}
